# api/docs/swagger.yaml
definitions:
...
  memory.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      updated_at:
        type: string
      writer_id:
        type: string
    type: object
  memory.CommentCreateBody:
    properties:
      content:
        type: string
      memory_id:
        type: string
    type: object
  memory.CommentUpdateBody:
    properties:
      content:
        type: string
    type: object
...
  memory.Media:
    properties:
      created_at:
        type: string
      id:
        type: string
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  memory.MediaCreateBody:
    properties:
      memory_id:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for Time Capsule resources
  title: Time Capsule API
  version: "1.0"
paths:
  /comments:
    delete:
      consumes:
      - application/json
      description: Delete a comment by ID
      parameters:
      - description: Comment ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
    get:
      consumes:
      - application/json
      description: Get all comments for a memory
      parameters:
      - description: Memory ID
        in: query
        name: memory_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Comment'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a new comment on a memory
      parameters:
      - description: Comment Create Body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/memory.CommentCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Comment created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update the content of an existing comment
      parameters:
      - description: Comment ID
        in: query
        name: id
        required: true
        type: string
      - description: Comment Update Body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/memory.CommentUpdateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - comments
  /media:
    delete:
      consumes:
      - application/json
      description: Delete a media entry by ID
      parameters:
      - description: Media ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete media
      tags:
      - media
    get:
      consumes:
      - application/json
      description: Get a media entry by ID
      parameters:
      - description: Memory ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/memory.Media'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get media
      tags:
      - media
    post:
      consumes:
      - application/json
      description: Create a new media entry
      parameters:
      - description: Media Create Body
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/memory.MediaCreateBody'
      produces:
      - application/json
      responses:
        "200":
          description: Media created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create media
      tags:
      - media
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"